Refactor loss placement for OD
